added middleware for admin
