+++ server/api/index.js
@@ -8,39 +8,7 @@ const jwt = require('jsonwebtoken');
 const prisma = require('../db/client');
 const { JWT_SECRET } = process.env;
 
-// set `req.user` if possible
-apiRouter.use(async (req, res, next) => {
-    const prefix = 'Bearer ';
-    const auth = req.header('Authorization');
 
-    if (!auth) {
-        // nothing to see here
-        next();
-    } else if (auth.startsWith(prefix)) {
-        const token = auth.slice(prefix.length);
-
-        try {
-            const { id } = jwt.verify(token, JWT_SECRET);
-
-            if (id) {
-                req.user = await prisma.user.findUnique({ where: { id: id } });
-                next();
-            } else {
-                next({
-                    name: 'AuthorizationHeaderError',
-                    message: 'Authorization token malformed',
-                });
-            }
-        } catch ({ name, message }) {
-            next({ name, message });
-        }
-    } else {
-        next({
-            name: 'AuthorizationHeaderError',
-            message: `Authorization token must start with ${prefix}`,
-        });
-    }
-});
 
 apiRouter.use((req, res, next) => {
     if (req.user) {
